requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
psycopg[binary]==3.2.10
supabase==2.0.0
flask-cors==4.0.0